gateway/python/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
slowapi>=0.1.9
python-dotenv>=1.0.1
aiofiles>=24.1.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.27.0
